# backend/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Classroom Supplies Usage & Replenishment Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /supplies:
    get:
      operationId: listSupplies
      tags: [supplies]
      summary: List all supplies
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of supplies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Supply"
    post:
      operationId: createSupply
      tags: [supplies]
      summary: Create a supply item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplyInput"
      responses:
        "201":
          description: Created supply
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supply"
...
  /supplies/{id}:
    get:
      operationId: getSupply
      tags: [supplies]
      summary: Get a supply by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Supply details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supply"
        "404":
          description: Not found
    patch:
      operationId: updateSupply
      tags: [supplies]
      summary: Update a supply
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplyUpdate"
      responses:
        "200":
          description: Updated supply
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supply"
        "404":
          description: Not found
    delete:
      operationId: deleteSupply
      tags: [supplies]
      summary: Delete a supply
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /classrooms:
    get:
      operationId: listClassrooms
      tags: [classrooms]
      summary: List all classrooms
      responses:
        "200":
          description: List of classrooms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Classroom"
    post:
      operationId: createClassroom
      tags: [classrooms]
      summary: Create a classroom
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClassroomInput"
      responses:
        "201":
          description: Created classroom
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Classroom"
...
  /classrooms/{id}:
    get:
      operationId: getClassroom
      tags: [classrooms]
      summary: Get a classroom by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Classroom
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Classroom"
        "404":
          description: Not found
    patch:
      operationId: updateClassroom
      tags: [classrooms]
      summary: Update a classroom
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClassroomUpdate"
      responses:
        "200":
          description: Updated classroom
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Classroom"
        "404":
          description: Not found
    delete:
      operationId: deleteClassroom
      tags: [classrooms]
      summary: Delete a classroom
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Successfully logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Invalid username or password
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get the current logged-in user profile
      responses:
        "200":
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register a new teacher user account (requires admin approval)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Successfully registered. Pending approval.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Validation error or username already taken.
...
  /classroom-supplies:
    get:
      operationId: listClassroomSupplies
      tags: [classrooms]
      summary: List supplies allocated to a classroom
      parameters:
        - name: classroomId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of allocated supplies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClassroomSupply"
...
components:
  schemas:
...
    Supply:
      type: object
      required: [id, name, category, quantity, unit, reorderThreshold, reorderQuantity, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        quantity:
          type: number
        unit:
          type: string
        reorderThreshold:
          type: number
        reorderQuantity:
          type: number
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    SupplyInput:
      type: object
      required: [name, category, quantity, unit, reorderThreshold, reorderQuantity]
      properties:
        name:
          type: string
          minLength: 1
        category:
          type: string
          minLength: 1
        quantity:
          type: number
          minimum: 0
        unit:
          type: string
          minLength: 1
        reorderThreshold:
          type: number
          minimum: 0
        reorderQuantity:
          type: number
          minimum: 1
        notes:
          type: string
    SupplyUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        category:
          type: string
          minLength: 1
        quantity:
          type: number
          minimum: 0
        unit:
          type: string
          minLength: 1
        reorderThreshold:
          type: number
          minimum: 0
        reorderQuantity:
          type: number
          minimum: 1
        notes:
          type: string
...
    Classroom:
      type: object
      required: [id, name, grade, teacher, roomNumber, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        grade:
          type: string
        teacher:
          type: string
        roomNumber:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    ClassroomInput:
      type: object
      required: [name, grade, teacher, roomNumber, teacherUsername, teacherPassword]
      properties:
        name:
          type: string
          minLength: 1
        grade:
          type: string
          minLength: 1
        teacher:
          type: string
          minLength: 1
        roomNumber:
          type: string
          minLength: 1
        notes:
          type: string
        teacherUsername:
          type: string
          minLength: 3
        teacherPassword:
          type: string
          minLength: 6
    ClassroomUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        grade:
          type: string
          minLength: 1
        teacher:
          type: string
          minLength: 1
        roomNumber:
          type: string
          minLength: 1
        notes:
          type: string
        teacherUsername:
          type: string
          minLength: 3
        teacherPassword:
          type: string
          minLength: 6
        approved:
          type: boolean
...
    User:
      type: object
      required: [id, username, name, role, approved]
      properties:
        id:
          type: integer
        username:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, teacher]
        approved:
          type: boolean
        classroomId:
          type: ["integer", "null"]
    LoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 1
...
    ClassroomSupply:
      type: object
      required: [id, classroomId, classroomName, supplyId, supplyName, quantity, unit]
      properties:
        id:
          type: integer
        classroomId:
          type: integer
        classroomName:
          type: string
        supplyId:
          type: integer
        supplyName:
          type: string
        quantity:
          type: number
        unit:
          type: string
...
    RegisterInput:
      type: object
      required: [username, name, password, classroomId]
      properties:
        username:
          type: string
          minLength: 1
        name:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 6
        classroomId:
          type: integer
...